+++ changelog.md
@@ -1,5 +1,17 @@
 Thank you for choosing to test Artona, please join our discord server in order to get support or request new sites to be added.
 
+Beta 5:
+- Made the color changes reflect on Artona popup after a page reload
+- Better Twitter.com integration
+- Fixed the JSON error on the form
+
+Beta 4:
+- Fixed Feedback form
+
+Beta 3:
+- Added an option to request sites to be added
+- Improved Youtube AdBlocking
+
 Beta 2:
 - Added Twitter.com support (Partial)
 - Added AdBlocker for Youtube.com